+++ kinto_signer/listeners.py
@@ -39,8 +39,6 @@ def pick_resource_and_signer(request, resources, bucket_id, collection_id):
                                   bucket_id=bucket_id,
                                   id=collection_id)
 
-    settings = request.registry.settings
-
     resource = signer = None
 
     # Review might have been configured explictly for this collection,
@@ -56,17 +54,6 @@ def pick_resource_and_signer(request, resources, bucket_id, collection_id):
         resource["destination"]["collection"] = collection_id
         if "preview" in resource:
             resource["preview"]["collection"] = collection_id
-
-        # Look-up if a setting overrides a global one here.
-        for setting in REVIEW_SETTINGS:
-            setting_key = "signer.%s.%s.%s" % (bucket_id, collection_id, setting)
-            if setting_key in settings:
-                resource[setting] = settings[setting_key]
-            else:  # pragma: no cover
-                # Deprecated underscore separation
-                setting_key = "signer.%s_%s.%s" % (bucket_id, collection_id, setting)
-                if setting_key in settings:
-                    resource[setting] = settings[setting_key]
 
     if collection_key in request.registry.signers:
         signer = request.registry.signers[collection_key]